requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.1
pydantic>=2.7.0
pydantic-settings==2.9.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
requests==2.31.0
beautifulsoup4==4.12.3
pytest-asyncio==0.21.1 